breathing exercise styles added to seperate file
